% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{aam_palettes}
\alias{aam_palettes}
\title{AAM Palettes}
\usage{
aam_palettes(
  name,
  n,
  all_palettes = all_aam_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{Which palette? One of "summer" or "winter"}

\item{n}{How many colors. Can be missing.}

\item{all_palettes}{List of palettes to consider, default is fine.}

\item{type}{One of c("discrete", "continuous")}
}
\value{
a structured palette S3 object
}
\description{
AAM Palettes
}
\examples{
aam_palettes("summer")
\dontrun{
require(ggplot2)
data(iris)
}
}
