% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{blue_theme}
\alias{blue_theme}
\title{Blue theme}
\usage{
blue_theme()
}
\value{
Whatever the ggplot object is
}
\description{
Taken from: https://emanuelaf.github.io/own-ggplot-theme.html
}
\examples{
\dontrun{
require(ggplot2)
data(iris)
ggplot(iris, aes(x = Petal.Length)) +
geom_histogram() +
blue_theme()
}
}
