% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_colour_aam_c}
\alias{scale_colour_aam_c}
\title{Continuous colors from aamColors}
\usage{
scale_colour_aam_c(name)
}
\arguments{
\item{name}{Palette to use.}
}
\value{
ggplot2 object
}
\description{
Continuous colors from aamColors
}
\examples{
\dontrun{
require(ggplot2)
data(iris)
}
}
