% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_colour_aam_d}
\alias{scale_colour_aam_d}
\title{Discrete colors from aamColors}
\usage{
scale_colour_aam_d(name)
}
\arguments{
\item{name}{Which palette to use.}
}
\value{
ggplot2 object
}
\description{
Discrete colors from aamColors
}
\examples{
\dontrun{
require(ggplot2)
data(iris)
ggplot(iris, aes(x = Petal.Length)) +
geom_histogram(aes(color = Species)) +
scale_color_aam_d("summer")
}
}
