% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_aam_c}
\alias{scale_fill_aam_c}
\title{Title}
\usage{
scale_fill_aam_c(name)
}
\arguments{
\item{name}{}
}
\value{
ggplot2 object
}
\description{
Title
}
\examples{
\dontrun{
require(ggplot2)
data(iris)
}
}
