% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_aam_d}
\alias{scale_fill_aam_d}
\title{Discrete fills from aamColors}
\usage{
scale_fill_aam_d(name)
}
\arguments{
\item{name}{Which palette to use.}
}
\value{
ggplot2 object
}
\description{
Discrete fills from aamColors
}
\examples{
\dontrun{
require(ggplot2)
data(iris)
#' ggplot(iris, aes(x = Petal.Length)) +
geom_histogram(aes(fill = Species)) +
scale_fill_aam_d("summer")
}
}
