% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_summer_day}
\alias{theme_summer_day}
\title{Summer day theme}
\usage{
theme_summer_day()
}
\value{
Ggplot object
}
\description{
Summer day theme
}
\examples{
\dontrun{
require(ggplot2)
data(iris)
ggplot(iris, aes(x = Petal.Length)) +
geom_histogram() +
theme_summer_day()
}
}
